frontend: added charts.js
got the actual chart to work
#TODO
Find a way to cache data properly, as it is lost as soon as the page refreshes.
Plug that data into the graph, then as soon as the other data comes through, replace the data with the new one.
Fix the graph being responsive, make it fit the container.